paso a TypeScript de los componentes y creación del router
