+++ package.json
@@ -1,0 +1,9 @@
+{
+  "dependencies": {
+    "react-router-dom": "^7.2.0",
+    "typescript": "^5.7.3"
+  },
+  "devDependencies": {
+    "@types/node": "^22.13.4"
+  }
+}
