+++ webapp/package.json
@@ -10,8 +10,8 @@
     "@testing-library/react": "^14.1.2",
     "@testing-library/user-event": "^14.5.2",
     "axios": "^1.7.9",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-scripts": "5.0.1",
     "react-simple-typewriter": "^5.0.1",
     "web-vitals": "^3.5.1"
@@ -43,6 +43,9 @@
     ]
   },
   "devDependencies": {
+    "@types/node": "^22.13.4",
+    "@types/react": "^19.0.10",
+    "@types/react-dom": "^19.0.4",
     "axios-mock-adapter": "^2.1.0",
     "expect-puppeteer": "^9.0.2",
     "jest": "^29.7.0",
