+++ webapp/src/components/ErrorInterface.ts
@@ -1,0 +1,3 @@
+export interface ErrorResponse {
+    error: string;
+}
