+++ webapp/src/react-app-env.d.ts
@@ -1,0 +1,5 @@
+declare namespace NodeJS {
+    interface ProcessEnv {
+        REACT_APP_API_ENDPOINT: string;
+    }
+}
